feat : 관리자 게시글 목록 조회, 상세 조회, 삭제 기능 추가
관리자 페이지에서 사용할 게시글 목록 조회, 게시글 상세 조회, 게시글 삭제 기능 구현

+++ src/main/java/com/yht/exerciseassist/domain/post/service/PostService.java
@@ -158,16 +158,21 @@ public class PostService {
         return new ResponseResult<>(HttpStatus.OK.value(), post.getTitle());
     }
 
-    public ResponseResult<Long> deletePost(Long postId) throws IOException {
+    public ResponseResult<Long> deletePost(Long postId) throws IOException, IllegalAccessException {
         Post postById = postRepository.findNotDeletedById(postId)
                 .orElseThrow(() -> new EntityNotFoundException(ErrorCode.NOT_FOUND_EXCEPTION_POST.getMessage()));
 
-        String localTime = LocalDateTime.now().format(DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm"));
+        if ((Objects.equals(postById.getPostWriter().getUsername(), SecurityUtil.getCurrentUsername()))||(SecurityUtil.getMemberRole().equals("ADMIN"))) {
 
-        postById.getDateTime().canceledAtUpdate();
-        commentRepository.deleteCommentByPostId(localTime, postById.getId());
-        mediaService.deletePostMedia(postById.getId());
-        likeCountRepository.deleteAllByPost(postById);
+            String localTime = LocalDateTime.now().format(DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm"));
+
+            postById.getDateTime().canceledAtUpdate();
+            commentRepository.deleteCommentByPostId(localTime, postById.getId());
+            mediaService.deletePostMedia(postById.getId());
+            likeCountRepository.deleteAllByPost(postById);
+        } else {
+            throw new IllegalAccessException(ErrorCode.NO_MATCHED_COMMENTID.getMessage());
+        }
 
         log.info("username : {}, {}번 게시글 삭제 완료", SecurityUtil.getCurrentUsername(), postById.getId());
         return new ResponseResult<>(HttpStatus.OK.value(), postById.getId());
